+++ app/Client/Mapper/AdSelect/EventMapper.php
@@ -22,8 +22,8 @@ declare(strict_types = 1);
 
 namespace Adshares\Adserver\Client\Mapper\AdSelect;
 
+use Adshares\Adserver\Client\Mapper\AbstractFilterMapper;
 use stdClass;
-use function substr;
 
 class EventMapper
 {
@@ -46,34 +46,12 @@ class EventMapper
         return $mappedEvent;
     }
 
-    private static function normalizeKeywords(?array $keywords = []): array
-    {
-        $mappedKeywords = [];
-
-        foreach ($keywords as $keyword) {
-            $lastOccurrence = strrpos($keyword, ':');
-
-            if ($lastOccurrence === false) {
-                $mappedKeywords[$keyword] = 1;
-
-                continue;
-            }
-
-            $key = substr($keyword, 0, $lastOccurrence);
-            $value = substr($keyword, $lastOccurrence + 1);
-
-            $mappedKeywords[$key] = $value;
-        }
-
-        return $mappedKeywords;
-    }
-
     private static function getNormalizedKeywordsFromEvent($event): ?array
     {
         $keywords = null;
         $eventContext = $event['context'];
         if (is_object($eventContext) && property_exists($eventContext, 'site')) {
-            $keywords = self::normalizeKeywords((array)$eventContext->site->keywords);
+            $keywords = AbstractFilterMapper::generateNestedStructure(['site'=>(array)$eventContext->site->keywords]);
         }
 
         return $keywords;