+++ app/Console/Commands/AdPayEventExportCommand.php
@@ -30,15 +30,18 @@ use Adshares\Common\Exception\Exception;
 use Adshares\Demand\Application\Service\AdPay;
 use Adshares\Supply\Application\Dto\ImpressionContextException;
 use Adshares\Supply\Application\Dto\UserContext;
+use DateTime;
 use Illuminate\Console\Command;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Facades\Log;
+use Symfony\Component\Console\Command\LockableTrait;
 use function json_encode;
 use function sprintf;
 
 class AdPayEventExportCommand extends Command
 {
     use LineFormatterTrait;
+    use LockableTrait;
 
     private const EVENTS_BUNDLE_MAXIMAL_SIZE = 100;
 
@@ -48,8 +51,14 @@ class AdPayEventExportCommand extends Command
 
     public function handle(AdPay $adPay, AdUser $adUser): void
     {
+        if (!$this->lock()) {
+            $this->info('[AdPayEventExport] Command '.$this->signature.' already running.');
+
+            return;
+        }
+
         $timeStart = microtime(true);
-        $this->info('[AdPayEventExport] Start command '.$this->signature);
+        $this->info('Start command '.$this->signature);
 
         $eventIdFirst = Config::fetchInt(Config::ADPAY_LAST_EXPORTED_EVENT_ID) + 1;
 
@@ -78,6 +87,7 @@ class AdPayEventExportCommand extends Command
     private function fetchEventsToExport(int $eventIdFirst): Collection
     {
         return EventLog::where('id', '>=', $eventIdFirst)
+            ->where('created_at', '<=', new DateTime('-10 minutes'))
             ->orderBy('id')
             ->limit(self::EVENTS_BUNDLE_MAXIMAL_SIZE)
             ->get();
@@ -93,11 +103,7 @@ class AdPayEventExportCommand extends Command
             }
 
             try {
-                $userContext = $this->userContext($adUser, $event);
-
-                $event->human_score = $userContext->humanScore();
-                $event->our_userdata = $userContext->keywords();
-
+                $event->updateWithUserContext($this->userContext($adUser, $event));
                 $event->save();
             } catch (ImpressionContextException $e) {
                 Log::error(
@@ -116,25 +122,26 @@ class AdPayEventExportCommand extends Command
     {
         static $userInfoCache = [];
 
-        $trackingId = $event->impressionContext()->trackingId();
+        $impressionContext = $event->impressionContextForAdUserQuery();
+        $trackingId = $impressionContext->trackingId();
 
         if (isset($userInfoCache[$trackingId])) {
             return $userInfoCache[$trackingId];
         }
 
-        $userContext = $adUser->getUserContext($event->impressionContext());
+        $userContext = $adUser->getUserContext($impressionContext);
 
         if ($userContext->humanScore() > AdUser::HUMAN_SCORE_MINIMUM) {
             $userInfoCache[$trackingId] = $userContext;
         }
 
         Log::debug(sprintf(
-            '%s {"userInfoCache": "MISS", "humanScore":%s, "event": %s, "userId": %s "context": %s}',
+            '%s {"userInfoCache":"MISS","humanScore":%s,"event":%s,"trackingId":%s,"context": %s}',
             __FUNCTION__,
             $userContext->humanScore(),
             $event->id,
-            $event->user_id,
-            json_encode($userContext->toArray())
+            $event->tracking_id,
+            json_encode($userContext->toArray()) ?: 'null'
         ));
 
         return $userContext;