+++ app/Models/Campaign.php
@@ -46,8 +46,8 @@ use function hex2bin;
  * @property int deleted_at
  * @property int user_id
  * @property string landing_url
- * @property \DateTime time_start
- * @property \DateTime time_end
+ * @property DateTime time_start
+ * @property DateTime time_end
  * @property int status
  * @property string name
  * @property array|null|string strategy_name
@@ -215,7 +215,7 @@ class Campaign extends Model
         return $this->banners;
     }
 
-    public function getTargetingAttribute()
+    public function getTargetingAttribute(): array
     {
         return [
             "requires" => $this->targeting_requires,
@@ -223,7 +223,7 @@ class Campaign extends Model
         ];
     }
 
-    public function setBasicInformationAttribute(array $value)
+    public function setBasicInformationAttribute(array $value): void
     {
         $this->status = $value["status"];
         $this->name = $value["name"];
@@ -238,7 +238,7 @@ class Campaign extends Model
         $this->time_end = $value["date_end"] ?? null;
     }
 
-    public function getBasicInformationAttribute()
+    public function getBasicInformationAttribute(): array
     {
         return [
             "status" => $this->status,
@@ -295,6 +295,7 @@ class Campaign extends Model
         $blockedAmount = abs(UserLedgerEntry::fetchBlockedAmountByUserId($this->user_id));
         if ($amount <= $blockedAmount) {
             Log::info(sprintf('Hold the blockade %d, while total budget is %d', $blockedAmount, $amount));
+
             return;
         }