+++ app/Models/EventLog.php
@@ -25,8 +25,10 @@ use Adshares\Adserver\Models\Traits\AccountAddress;
 use Adshares\Adserver\Models\Traits\AutomateMutators;
 use Adshares\Adserver\Models\Traits\BinHex;
 use Adshares\Adserver\Models\Traits\JsonValue;
+use Adshares\Common\Domain\ValueObject\Uuid;
 use Adshares\Common\Exception\RuntimeException;
 use Adshares\Supply\Application\Dto\ImpressionContext;
+use Adshares\Supply\Application\Dto\UserContext;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
@@ -43,6 +45,7 @@ use function sprintf;
  * @property string case_id
  * @property string event_id
  * @property string user_id
+ * @property string tracking_id
  * @property string banner_id
  * @property string publisher_id
  * @property string advertiser_id
@@ -90,6 +93,7 @@ class EventLog extends Model
         'case_id',
         'event_id',
         'user_id',
+        'tracking_id',
         'banner_id',
         'zone_id',
         'publisher_id',
@@ -129,6 +133,7 @@ class EventLog extends Model
         'case_id' => 'BinHex',
         'event_id' => 'BinHex',
         'user_id' => 'BinHex',
+        'tracking_id' => 'BinHex',
         'zone_id' => 'BinHex',
         'banner_id' => 'BinHex',
         'publisher_id' => 'BinHex',
@@ -186,7 +191,7 @@ class EventLog extends Model
         $log->case_id = $caseId;
         $log->event_id = $eventId;
         $log->banner_id = $bannerId;
-        $log->user_id = $trackingId;
+        $log->tracking_id = $trackingId;
         $log->zone_id = $zoneId;
         $log->publisher_id = $publisherId;
         $log->campaign_id = $campaignId;
@@ -226,7 +231,7 @@ class EventLog extends Model
         return $this->belongsTo(Payment::class);
     }
 
-    public function impressionContext(): ImpressionContext
+    public function impressionContextForAdUserQuery(): ImpressionContext
     {
         $headersArray = get_object_vars($this->headers);
 
@@ -239,16 +244,16 @@ class EventLog extends Model
         $ua = $userAgentList[0] ?? '';
 
         try {
-            $userId = Utils::trackingIdFromBinUserId(hex2bin($this->user_id));
+            $trackingId = Utils::base64UrlEncodeWithChecksumFromBinUuidString(hex2bin($this->tracking_id));
         } catch (RuntimeException $e) {
-            Log::warning(sprintf('%s %s', $e->getMessage(), $this->user_id));
-            $userId = '';
+            Log::warning(sprintf('%s %s', $e->getMessage(), $this->tracking_id));
+            $trackingId = '';
         }
 
         return new ImpressionContext(
             ['domain' => $domain, 'page' => $domain],
             ['ip' => $ip, 'ua' => $ua],
-            ['uid' => $userId]
+            ['tid' => $trackingId]
         );
     }
 
@@ -257,5 +262,15 @@ class EventLog extends Model
         self::where('case_id', hex2bin($caseId))
             ->where('event_type', self::TYPE_VIEW)
             ->update(['is_view_clicked' => 1]);
+    }
+
+    public function updateWithUserContext(UserContext $userContext): void
+    {
+        $userId = $userContext->userId();
+        if ($userId) {
+            $this->user_id = Uuid::fromString($userId)->hex();
+        }
+        $this->human_score = $userContext->humanScore();
+        $this->our_userdata = $userContext->keywords();
     }
 }