+++ app/Models/Zone.php
@@ -28,10 +28,8 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Log;
 use function array_map;
-use function count;
-use function GuzzleHttp\json_encode;
+use function array_unique;
 use function hex2bin;
 
 /**
@@ -130,6 +128,7 @@ HTML;
         'size',
         'status',
         'type',
+        'uuid'
     ];
 
     protected $appends = [
@@ -155,24 +154,14 @@ HTML;
 
     public static function findByPublicIds(array $publicIds): Collection
     {
-        $binPublicIds = array_map(
+        $binUniquePublicIds = array_unique(array_map(
             function (string $item) {
                 return hex2bin($item);
             },
             $publicIds
-        );
+        ));
 
-        $zones = self::whereIn('uuid', $binPublicIds)->get();
-
-        if (count($zones) !== count($binPublicIds)) {
-            Log::warning(sprintf(
-                'Missing zones. {"ids":%s,"zones":%s}',
-                json_encode($publicIds),
-                json_encode($zones->pluck(['id', 'width', 'height'])->toArray())
-            ));
-        }
-
-        return $zones;
+        return self::whereIn('uuid', $binUniquePublicIds)->get();
     }
 
     public static function fetchPublisherPublicIdByPublicId(string $publicId): string