+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.4] - 2019-04-24
+### Added
+- Command execution lock
+### Improved
+- Inventory export (keyword mapping)
+
+## [0.6.3] - 2019-04-23
+### Fixed
+- Event export
+
 ## [0.6.2] - 2019-04-19
 ### Fixed
 - Inventory import new banners - could not map to demand id which not exist
@@ -83,7 +94,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v0.6.2...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v0.6.4...develop
+[0.6.4]: https://github.com/adshares/adserver/compare/v0.6.3...v0.6.4
+[0.6.3]: https://github.com/adshares/adserver/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/adshares/adserver/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/adshares/adserver/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/adshares/adserver/compare/v0.5.4...v0.6.0